\docType{data}
\name{exampleDataset}
\alias{exampleDataset}
\title{Example Dataset}
\description{
Creates random input data myY and myX
}
\details{
The variables included in the dataset are: \itemize{
\item\code{myY} A matrix object; The number of columns is
one; The number of rows depends on the data \item\code{myX}
A matrix object; The number of rows is the same as that of
\code{myY}; The number of columns depends on the data }
}

